backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.2.1
requests==2.32.5
motor==3.3.1
pymongo==4.5.0
groq
opencv-python
pillow
emergentintegrations
treys
deuces
numpy
# Librerie per Client Desktop (Windows)
mss
pygetwindow
PyQt5
